requirements.txt
PyAutoGUI>=0.9.52
pytesseract>=0.3.7
Pillow>=8.1.0
googletrans>=4.0.0rc1
SQLAlchemy>=1.3.22
opencv-python>=4.5.1.48
numpy>=1.19.5